fix keyboard overthe screen and create personal component
